+++ app/models/social/social_profile.rb
@@ -6,7 +6,7 @@ class SocialProfile < ActiveRecord::Base
   validates :age, numericality: {only_integer: true, less_than_or_equal_to: 120, allow_nil: true}
   validates :sex, inclusion: { in: %w(Male Female Other), allow_nil: true}
 
-  def self.locations_for_map(user=nil)
+  def self.locations_for_map(user)
     res = select(:latitude, :longitude).where(show_location: true)
     res = res.where.not(id: user.social_profile.id) if user and user.social_profile