+++ app/models/surveys/question_flow.rb
@@ -147,11 +147,7 @@ class QuestionFlow < ActiveRecord::Base
     update_attributes(tsorted_edges: nil, longest_time: nil, longest_path: nil)
   end
 
-
-
-
   private
-
 
 
 end