+++ config/application.rb
@@ -26,7 +26,6 @@ module OpenpprnBackend
     # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
     # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
     # config.i18n.default_locale = :de
-    config.layout = 'myapnea/myapnea'
 
     config.autoload_paths += Dir[Rails.root.join('app', 'models', '{**/}')]
   end