+++ config/initializers/forem.rb
@@ -2,7 +2,7 @@ require 'emoji'
 Forem.user_class = "User"
 Forem.main_layout = "myapnea/myapnea"
 Forem.admin_layout = "admin"
-Forem.logged_in_layout = "dashboard"
+Forem.logged_in_layout = "social"
 Forem.avatar_user_method = :photo_url
-Forem.email_from_address = "support@myapnea.org"
+Forem.email_from_address = "support@openpprn.org"
 Forem.per_page = 20