+++ db/migrate/20141002132006_update_social_profile.rb
@@ -1,9 +1,0 @@
-class UpdateSocialProfile < ActiveRecord::Migration
-  def change
-    # More options for visibility and privacy
-    # Forem, Research Topics, Map for community, outside.
-
-    add_column :social_profiles, :visible_to_community, :boolean, default: false, null: false
-    add_column :social_profiles, :visible_to_world, :boolean, default: false, null: false
-  end
-end