+++ test/controllers/forem/posts_controller_test.rb
@@ -1,6 +1,0 @@
-class Forem::PostsControllerTest < ActionController::TestCase
-
-
-
-
-end