+++ test/controllers/forem/topics_controller_test.rb
@@ -1,5 +1,0 @@
-class Forem::TopicsControllerTest < ActionController::TestCase
-
-
-
-end