+++ CHANGELOG.md
@@ -3,10 +3,9 @@
 ## R2
 
 ## Enhancements
-- Research Democracy - voting for research topics.
+- Research Democracy - submitting and voting for research topics.
 
 ## Fixes
-- home page learn and share and research boxes now have the same height.
 
 ## R1 - Social Functionality
 ### Enhancements